Utils methods messing in main Moved to separate iutil package
