+++ cmd/nem/main.go
@@ -9,14 +9,13 @@ import (
 	"os"
 	"time"
 
-	"encoding/hex"
-
 	"runtime"
 
 	"strings"
 
 	"math"
 
+	"github.com/nem-toochain/nem-toolchain/pkg/util"
 	"github.com/nem-toolchain/nem-toolchain/pkg/core"
 	"github.com/nem-toolchain/nem-toolchain/pkg/keypair"
 	"github.com/nem-toolchain/nem-toolchain/pkg/vanity"
@@ -103,7 +102,7 @@ func generateAction(c *cli.Context) error {
 
 	num := c.Uint("number")
 	for i := uint(0); i < num; i++ {
-		printAccountDetails(ch, keypair.Gen())
+		util.PrintAccountDetails(ch, keypair.Gen())
 		fmt.Println("----")
 	}
 
@@ -149,7 +148,7 @@ func vanityAction(c *cli.Context) error {
 				fmt.Print(".")
 			}
 		}()
-		rate := allCPUKeyPairsInSeconds()
+		rate := util.CPUKeyPairsInSeconds()
 		ticker.Stop()
 		fmt.Printf(" %v accounts/sec\n", math.Trunc(rate))
 
@@ -158,9 +157,9 @@ func vanityAction(c *cli.Context) error {
 			fmt.Printf("Specified search complexity: %v\n", math.Trunc(1.0/pbty))
 		}
 		fmt.Printf("Estimate search times: %v (50%%), %v (80%%), %v (99.9%%)\n",
-			printTimeInSeconds(numberOfKeyPairs(pbty, 0.5)/rate),
-			printTimeInSeconds(numberOfKeyPairs(pbty, 0.8)/rate),
-			printTimeInSeconds(numberOfKeyPairs(pbty, 0.99)/rate))
+			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.5)/rate),
+			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.8)/rate),
+			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.99)/rate))
 		fmt.Println("----")
 	}
 
@@ -173,7 +172,7 @@ func vanityAction(c *cli.Context) error {
 		if i != 0 {
 			fmt.Println("----")
 		}
-		printAccountDetails(ch, <-rs)
+		util.PrintAccountDetails(ch, <-rs)
 		go vanity.StartSearch(ch, sel, rs)
 	}
 
@@ -193,48 +192,4 @@ func chainGlobalOption(c *cli.Context) (core.Chain, error) {
 		return ch, fmt.Errorf("unknown chain '%v'", c.GlobalString("chain"))
 	}
 	return ch, nil
-}
-
-func allCPUKeyPairsInSeconds() (rate float64) {
-	res := make(chan int, runtime.NumCPU())
-	for i := 0; i < cap(res); i++ {
-		go func(res chan<- int) {
-			res <- countKeyPairs(3200)
-		}(res)
-	}
-	for i := 0; i < cap(res); i++ {
-		rate += float64(<-res) / 3.2
-	}
-	return
-}
-
-func countKeyPairs(milliseconds time.Duration) int {
-	timeout := time.After(time.Millisecond * milliseconds)
-	for count := 0; ; count++ {
-		keypair.Gen().Address(core.Mainnet)
-		select {
-		case <-timeout:
-			return count
-		default:
-			continue
-		}
-	}
-}
-
-func numberOfKeyPairs(pbty, prec float64) float64 {
-	return math.Log(1-prec) / math.Log(1-pbty)
-}
-
-func printTimeInSeconds(val float64) string {
-	val = 1e9 * math.Trunc(val)
-	if val >= math.MaxInt64 || math.IsInf(val, 0) {
-		return "Inf"
-	}
-	return time.Duration(val).String()
-}
-
-func printAccountDetails(chain core.Chain, pair keypair.KeyPair) {
-	fmt.Println("Address:", pair.Address(chain).PrettyString())
-	fmt.Println("Public key:", hex.EncodeToString(pair.Public))
-	fmt.Println("Private key:", hex.EncodeToString(pair.Private))
 }
